Updating to latest stable version.
This version adds a third option to rotate noise to spectra (currently the preferred method) which is less agressive than method 2. It applies a non-linear multiplication as a function of frequency, which gets stronger the further away you are from the central frequency. This can be tuned in the config file. If you want it to be weaker you should reduce the total frequency map range, usally by orders of 10. So 0.001, - 1 will have a stronger non-linear effect than 0.01 - 1 will, for example.